move enums to global.zhl
