+++ CMakeLists.txt
@@ -173,7 +173,6 @@ target_sources(
     EnemyShipIcons.cpp
     EnemyShipIcons.h
     EnumClassHash.h
-    Enums.h
     Equipment_Extend.cpp
     Equipment_Extend.h
     ErosionEffect.cpp
