+++ FTL_HyperspaceGCC.cbp
@@ -1097,9 +1097,6 @@
 		<Unit filename="EnumClassHash.h">
 			<Option virtualFolder="PAL/" />
 		</Unit>
-		<Unit filename="Enums.h">
-			<Option virtualFolder="Core/" />
-		</Unit>
 		<Unit filename="Equipment_Extend.cpp">
 			<Option virtualFolder="Core/Equipment/" />
 		</Unit>
